Organize tests for merchant relationship endpoints into separate files
